requirements.txt
langchain==0.3.6
langchain-anthropic==0.2.4
langchain-core==0.3.15
python-decouple==3.8